Add /_sass, Add /assets, Catch-up to Local
Catching this up to the current stage of local development.

Created basic stylic for the theme, heavily based on Jekyll's "minima" default theme. This is handled by the /_sass and /assets directories.

Footer, head, and header have been updated.

Also added assets/minima-social-icons.svg to .gitignore in order to prevent accidentally commiting it when switching from the local environment