Portfolio and portfolio_items Collection
Created a new layout, portfolio-home.html, which inherits from the default.html layout. It also includes a newly created portfolio-header.html include, which is used to link to the portfolio-related pages (specifically through collections)

Also updated the header.hml include so that "Home" displays first. NOTE: In the future, it would be better to implement some sort of "order" page variable and sort by that on-the-fly as needed. I plan to do this with portfolio_items as well

Also added some new style-related options in _layout.scss related to the portfolio.