Portfolio Building
WARNING: _includes/header.html differs from the local instance in that the "Portfolio" link is not displayed in the nav, as I don't feel it's ready for live display.

Made a framework for the creation of specific project pages (hardware and software) and "experience pages" (such as a resume).
Created includes for displaying a type of page nav on a per-page basis dependent on properties related to a new collection called page_types

Deleted the portfolio-home.html layout, as I find it smarter to use includes combined with the new "backend tagging system".