Remove "seo" Tag from Comments in head.html and contact.html
The "head" include and the "contact" layout had recently been changed to comment out the "{%- seo -%}" content, however I overlooked the fact that HTML comments are parsed AFTER Jekyll generation and not BEFORE.

As a result, my changes in the earlier commit did not actually fix the issue surrounding removal of SEO content. After pushing this commit to the master branch, I am hopeful that local builds of the site from the ryanbranch/ryanbranch.github.io repository will be successful.