Minor Formatting Changes for Blogposts
Made one additional change to code syntax highlighting, as an extension of the previous commit.

Also extended the .post-content section of _layout.scss by adding a section specifically for <ul> amd <ol> tags, which applies the padding-left style with a value of 30 pixels. This has "stacking" effects, so not only the first but also all subsequent sections will experience further indent depending on their level of list depth.