Update post.html Layout
Modified the post-meta paragraph section in the post.html Layout, used by all blogpost pages.

Specifically, included the text "Written by {{ page.author }} on " before the date text below the blogpost title and subtitle. If no page.author variable is present, the generator will default to site.author instead.

The purpose of this commit is to address the following issue: #9