Update portfolio-table.html Layout
NOTE: This commit has the goal of addressing Issue #4 but no further action should be taken until the live site is rebuilt and tested.

Updated the portfolio-table Layout to prevent the specific software (and hardware) projects from being listed, thus making the page significantly less crowded.

In the future if changes are made such that a large amount of software project content exists and should be linked to, then the state directly before this commit can be used as a point of reference.