Update _config.yml
I only today realized that the site.title variable with the jekyanbranch theme _config.yml was still set to "using-local-overriding-minima".

Knowing that other sites implement this theme, there is no valid reason to remove the configuration file completely. So instead, this commit serves to standardize the file such that it is more appropriate for the individual theme itself. Variables such as site.description and site.github_username are okay to define because they

1. Are present within Jekyll (Minima) by default
2. Will be overriden by any equally-named site variable in the _config.yml of any website implementing jekyanbranch