Attempt to Fix Blogpost Rendering
The previous commit made the new <hr> styles functional.

This commit now modifies the hr-post-header style to use a margin-top value of -10px.

Within the post.html Layout, it also moves the hr-hidden-br-half instance up to before the page.author if statements. Previously, this linebreak was mistakenly inserted into middle of the line with author and date information.