Modify footer.html Include and CSS Generation
Changed the copyright statement in the footer to be a two-column, one-row table. The right cell holds the copyright statement and the left cell holds an added message.

Also added new CSS classes (in _base.scss) for <table>, <tr>, <th>, and <td> elements. This will allow for more well-organized display of information generated from the footer.html Include