Fix Table in footer.html Include
NOTE: In the future, inline styling of <div> elements within footer.html should be modified to instead reference a (generated) CSS document.

In the previous commit, changes were made to the footer layout which included a new table. This was intended to have transparent cell borders, but a CSS generation issue led to the styles not being properly applied.

To simplify the problem, I have taken a "<div> table" approach and used inline styles within the footer.html file, as opposed to any SCSS file.